requirements.txt
Django==1.7.4
django-debug-toolbar==1.4
graphql-core==0.4.12.1
